requirements.txt
wxPython
psutil
Pillow
